requirements.txt
gunicorn==22.0.0
agno==2.3.26
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
backports-datetime-fromisoformat==2.0.3
blinker==1.9.0
cachetools==6.2.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
distro==1.9.0
docstring_parser==0.17.0
environs==14.5.0
exceptiongroup==1.3.1
fastapi==0.128.0
fastapi_cors==0.0.6
gitdb==4.0.12
GitPython==3.1.46
greenlet==3.3.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
hyperframe==6.1.0
idna==3.11
Jinja2==3.1.6
jiter==0.12.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==4.2.1
mcp==1.25.0
mdurl==0.1.2
narwhals==2.15.0
numpy==2.2.6
openai==2.15.0
packaging==25.0
pandas==2.3.3
pillow==12.1.0
plotly==6.5.2
protobuf==6.33.4
pyarrow==22.0.0
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.21
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rich==14.2.0
rpds-py==0.30.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.45
sse-starlette==3.1.2
starlette==0.50.0
streamlit==1.53.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.4
tqdm==4.67.1
typer==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.40.0
watchdog==6.0.0
websockets==16.0
